Tap Highlighting Fails For Overflow Divs with --enable-accelerated-overflow-scroll
This CL modifies link highlighting to be compatible with the accelerated opt-in for overflow divs. It also fixes issues with highlights being incorrectly positioned when layers are promoted as a consequence of overflow div opt-in (e.g. in GMail). The CL also removes the need for special-case handling of boxshadows, and thus simplifies the link highlighting rect computation.

This CL (temporarily) re-baselines one test which fails under the revised code: the test does not correspond to any known bug, and is a corner case that (imo) should not stop this CL from proceeding given the relative necessity of the fixes this CL provides. The re-baselined test will be re-instated at a later time when a fix is found for it (a bug will be filed when this patch lands).

BUG=236905
R=jamesr@chromium.org, jchaffraix@chromium.org

Review URL: https://codereview.chromium.org/14803007

git-svn-id: svn://svn.chromium.org/blink/trunk@151573 bbb929c8-8fbe-4397-9dbb-9b2b20218538