Guard ResourceLoader::didReceiveData in error case.
A subresource can receive a body on a 404 if its call to
CachedResource::error() happens around a nested message loop. This can
cause a crash when data finally is received.

This is a regression brought about by refactoring, I thought I fixed
this in https://bugs.webkit.org/show_bug.cgi?id=103563 , but the fix
dissapeared.

Tested in a content_browsertest, see
https://codereview.chromium.org/15738007/ which can land after this is
gardened in.

BUG=241139
R=japhet

Review URL: https://chromiumcodereview.appspot.com/15725010

git-svn-id: svn://svn.chromium.org/blink/trunk@151609 bbb929c8-8fbe-4397-9dbb-9b2b20218538