Move core.gyp up one directory to Source/core/
This makes core match the build file layout in the rest of Blink.

Scripts previously living in core/core.gyp/scripts are now alongside
core/scripts/, which seems fine given that all of these scripts are for
build-time activities.

This is part one of a two-stage patch: core_tmp.gyp will be renamed to
core.gyp in a followup patch to keep the trybots happy.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/16357021

git-svn-id: svn://svn.chromium.org/blink/trunk@151856 bbb929c8-8fbe-4397-9dbb-9b2b20218538