Do not invoke new hover effects or clear existing hover effects on node deletion when cursor is invisible
If the mouse cursor is invisible and the currently
hovered node is deleted from the DOM, then there should
be no new hover effects invoked for any remaining nodes
in the page (i.e., those that happen to become positioned
underneath the last known mouse position as a result of
the deletion). Any existing hover effects for the ancestors
of the deleted node should be preserved.

BUG=240722
R=ojan@chromium.org, rbyers@chromium.org

Review URL: https://codereview.chromium.org/16285002

git-svn-id: svn://svn.chromium.org/blink/trunk@151861 bbb929c8-8fbe-4397-9dbb-9b2b20218538