Fix debug settings propagation from WebViewImpl to LayerTreeView
Currently, we propagate a bunch of debug settings from WebViewImpl
to LayerTreeView iff compositing is active. If compositing is not
active but m_layerTreeView exists, the value in layer tree view
becomes out of sync wrt that in WebViewImpl and is used next time
compositing is enabled.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/16229003

git-svn-id: svn://svn.chromium.org/blink/trunk@152104 bbb929c8-8fbe-4397-9dbb-9b2b20218538