Always return 0 or 1 from UnitBezier::solve() if the input is out of range
If the input to UnitBezier::solve() is outside the range [0, 1], then for some curve parameters and epsilon values, the existing logic attempts to solve for the y value. In other cases, it will simply return 0 for input values less than 0, or 1 for input values greater than 1.

This change makes the behavior more consistent by always returning 0 or 1 if the input is out of range. It also simplifies the code path a little.

This change also adds some tests for UnitBezier.

BUG=240653

Review URL: https://chromiumcodereview.appspot.com/15824002

git-svn-id: svn://svn.chromium.org/blink/trunk@152328 bbb929c8-8fbe-4397-9dbb-9b2b20218538