Mac: Fix scroll offset not being restored when using back-swipe gesture.
ScrollElasticityController::handleWheelEvent() is changed to only return
true for the PlatformWheelEventPhaseEnded case if a rubber band was actually
started (so that it doesn't return true when the start of rubber-banding
did not happen due to !shouldRubberBandInHorizontalDirection()).

EventHandler::handleWheelEvent() is changed to only set |wasScrolledByUser|
when returning true. Also, some clean up.

The underlying issue was caused by the fact that wasScrolledByUser() was
true in HistoryController::restoreScrollPositionAndViewState(), which caused
it to not restore the scroll state.

BUG=141130
TEST=On a Mac 10.7+, go to https://github.com/mojombo/jekyll/wiki/sites
and scroll down a bit. Click on a link. Gesture-scroll left via the
trackpad to trigger back navigation. The previous page should have the
same scroll position as before.

Review URL: https://chromiumcodereview.appspot.com/16336016

git-svn-id: svn://svn.chromium.org/blink/trunk@152374 bbb929c8-8fbe-4397-9dbb-9b2b20218538