HTMLDialogElement doesn't need to reattach manually
HTMLDialogElement::positionAndReattaach previously called reattach manually
which was not needed since we only need a style recalc to update the style of
the renderer and trigger a layout to put the dialog in the center.

This patch instead marks the dialog for an inline style change, since that's
really all that changed.

Review URL: https://chromiumcodereview.appspot.com/16997005

git-svn-id: svn://svn.chromium.org/blink/trunk@152397 bbb929c8-8fbe-4397-9dbb-9b2b20218538