Remove uses_expectation_file from results.html and use is_unexpected.
Blink always uses a TestExpectations file and now that we have
an is_unexpected bit in the results json, we don't need to redo the logic
for whether a failure is expected or not.

For ease of testing, I've still included that logic in the mock result code.

Review URL: https://chromiumcodereview.appspot.com/17084013

git-svn-id: svn://svn.chromium.org/blink/trunk@152483 bbb929c8-8fbe-4397-9dbb-9b2b20218538