Properly resolve named grid lines
This change adds the code to convert from a raw named grid line
(stored in RenderStyle) to its actual grid position.

The code is a pretty straightforward mapping from the specification
that builds on previous changes that store the right information in
RenderStyle. The rest of the code doesn't care about how we resolve
positions which is why the change is pretty small.

The named grid line resolution code has to be careful with resolving
sides properly and avoiding out-of-bounds access.

BUG=234192
TEST=fast/css-grid-layout/grid-item-named-grid-line-resolution.html
     fast/css-grid-layout/grid-item-negative-position-resolution.html

Review URL: https://chromiumcodereview.appspot.com/15023018

git-svn-id: svn://svn.chromium.org/blink/trunk@152659 bbb929c8-8fbe-4397-9dbb-9b2b20218538